pyproject.toml
[project]
name = "PwNo"
version = "0.1.5"
description = "Pwntools Extension that Just Works"
authors = [
    {name = "MuelNova", email = "n@ova.moe"},
]
dependencies = [
    "pwntools>=4.12.0",
    "pydantic>=2.8.2",
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "ruff>=0.5.5",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "C90"]
ignore = ["F405", "F403", "F841"]
